@@ -28,7 +28,7 @@ import (
 type CustomUmarshaller func(*elemental.Request) (elemental.Identifiable, error)
 
 // CustomMarshaller is the type of function use to create custom marshalling.
-type CustomMarshaller func(*elemental.Response, interface{}, error) ([]byte, error)
+type CustomMarshaller func(*elemental.Response, any, error) ([]byte, error)
 
 // RegisterProcessorOrDie will register the given Processor for the given
 // Identity and will exit in case of errors. This is just a helper for
@@ -223,7 +223,7 @@ func (b *server) RoutesInfo() map[int][]RouteInfo {
 	return buildVersionedRoutes(b.cfg.model.modelManagers, b.ProcessorForIdentity)
 }
 
-func (b *server) VersionsInfo() map[string]interface{} {
+func (b *server) VersionsInfo() map[string]any {
 
 	return b.cfg.meta.version
 }
