@@ -506,7 +506,7 @@ func TestWSPushSession_listen(t *testing.T) {
 			So(elemental.Decode(event.Encoding, eventData, &elemErr), ShouldBeNil)
 			So(elemErr.Title, ShouldEqual, "Bad request")
 			So(elemErr.Description, ShouldContainSubstring, "elemental: cannot declare an identity filter on \"identity-two\" as that was not declared in 'Identities'")
-			So(elemErr.Data, ShouldResemble, map[string]interface{}{
+			So(elemErr.Data, ShouldResemble, map[string]any{
 				"pushconfig": "filters",
 			})
 		})
@@ -577,7 +577,7 @@ func TestWSPushSession_listen(t *testing.T) {
 			So(elemental.Decode(event.Encoding, eventData, &elemErr), ShouldBeNil)
 			So(elemErr.Title, ShouldEqual, "Bad request")
 			So(elemErr.Description, ShouldContainSubstring, "elemental: unable to parse filter \"this-will-not-parse\": invalid operator")
-			So(elemErr.Data, ShouldResemble, map[string]interface{}{
+			So(elemErr.Data, ShouldResemble, map[string]any{
 				"pushconfig": "filters",
 			})
 		})
@@ -653,7 +653,7 @@ func TestWSPushSession_listen(t *testing.T) {
 			So(elemental.Decode(event.Encoding, eventData, &elemErr), ShouldBeNil)
 			So(elemErr.Title, ShouldEqual, "Bad request")
 			So(elemErr.Description, ShouldContainSubstring, "unsupported comparator: CONTAINS")
-			So(elemErr.Data, ShouldResemble, map[string]interface{}{
+			So(elemErr.Data, ShouldResemble, map[string]any{
 				"pushconfig": "filters",
 			})
 		})
